chore: 웹 설정 변경
